Enable Portal components to access Redux store
The Portal component renders itself directly with renderDOM which
means its children can no longer see the Redux store via context.

This feels like a broken hack, but basically grab the store from
context and then wrap the Portal's children in a react-redux Provider.

It should probably be the case that Modals do not directly interact
with the Redux store...