Run npm run build after Heroku build
We want the Heroku slug creation to include this build step (rather than
on npm start when a dyno starts up).  But if we use 'postinstall', this
also gets run when we do a 'npm install @pactcoffee/loggins' and we
don't want that.

So use the 'heroku-postinstall' hook so hopefully...

- npm run build happens in Heroku in the build stage
- npm install @pactcoffee/loggins works elsewhere